+++ example/lib/Database/Models/user.dart
@@ -1,7 +1,7 @@
 import 'task.dart';
 
 class User {
-  int? id;
+  String? id;
   String? name;
   String? email;
   String? phone;